requirements.txt
streamlit
python-docx
python-pptx
Pillow
matplotlib
pdfkit
reportlab